Adding design inspiration resources
Mobile Patterns, UI Garage, Mobbin, scrnshts, and dribbble

There weren't many resources in the design inspiration category so I added a few. All of which are free and somewhat different (and with different catalogs of content).

If you'd like me to remove any, I'd be happy to do so. I took the description from their pages directly.